Update README.md
LDR exhibits photoconductivity and changes 
its resistance depending on the light intensity falling on it. 
The resistance of the LDR increases when the intensity of 
the light decreases. For each object whose color is to be 
detected, the three colored LEDs (red, green and blue 
lights) are turned on sequentially and data is recorded by 
the LDR simultaneously. For example, during a scan of a 
red object, each LED is shining on the object which causes 
the reflected light from the object to fall on the LDR, and 
it is known that a red object will tend to reflect all the red 
light falling upon it while it absorbs the green and blue 
lights as shown in Figure-1. In such situation, minimal 
light intensity from the green and blue LEDs is reflected to 
the LDR, while a maximum light intensity is reflected 
from the red LED. Hence the LDR will have a larger 
resistance for the green and blue light scanning, while it 
has a minimum resistance for red light scanning.

The Arduino is controlling the LEDs. Each group 
of the LEDs has to have a variable resistor; to control the 5 
volts that the Arduino feeds and make sure all the LEDs 
has exactly the same brightness to avoid having major 
errors in the intensity which the LDR detects.A resistor should be used with the LDR to take 
the voltage difference in between. As the light intensity 
increases, the LDR’resistance will keep decreasing so that 
the voltage.
The Light from the LED is reflected by an object 
to the LDR. 
the LDR and the LEDs to avoid interference between the 
reflected light and the light from the LEDs directly. 
The LEDs system 
The voltage drops on the LDR when each of Red, 
Green and Blue LEDs are on separately is measured. Each 
group color of LEDs will turn on for 0.5 second and then 
turn off. When the LED color matches the color of the surface then it will have less voltage on that LED timing. 
For example, if the object is red, the red LED will have 
less voltage than the blue and green when it is on. 
If the object color is orange, the voltage will 
decrease in the same rate but slightly different from the 
Red.